bundled all Browser Dialog to one class + added fileorfolderbrowserdialog
