+++ pdf-tools/src/main/java/filetools/tiff/TiffTagAnalysis.java
@@ -17,7 +17,7 @@ public class TiffTagAnalysis {
 
 		try {
 
-			examinedFolder = utilities.FolderBrowserDialog.chooseFolder();
+			examinedFolder = utilities.BrowserDialogs.chooseFolder();
 
 			if (examinedFolder != null) {
