+++ pdf-tools/src/main/java/utilities/BrowserDialogs.java
@@ -5,7 +5,7 @@ import java.io.FileNotFoundException;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
 
-public class FolderBrowserDialog {
+public class BrowserDialogs {
 
 	/**
 	 * Chooses the folder which is examined via a simple folder browser Dialog
@@ -14,6 +14,7 @@ public class FolderBrowserDialog {
 	 *            not need any to begin with.
 	 * @return: string for folder path
 	 */
+	// TODO: search zip file
 
 	public static String chooseFolder() throws FileNotFoundException {
 		JFileChooser j = new JFileChooser();
@@ -27,7 +28,27 @@ public class FolderBrowserDialog {
 			return folder;
 		}
 	}
-	
+
+	/**
+	 * Chooses the file which is examined via a simple folder browser Dialog
+	 *
+	 * @param Does
+	 *            not need any to begin with.
+	 * @return: string for file path
+	 */
+	public static String chooseFile() throws FileNotFoundException {
+		JFileChooser j = new JFileChooser();
+		j.setFileSelectionMode(JFileChooser.FILES_ONLY);
+		j.showOpenDialog(j);
+		if (j.getSelectedFile() == null) {
+			JOptionPane.showMessageDialog(null, "No file was chosen");
+		} else {
+			String file = j.getSelectedFile().getPath();
+			return file;
+		}
+		return null;
+	}
+
 	public static String chooseFileOrFolder() throws FileNotFoundException {
 		JFileChooser j = new JFileChooser();
 		j.setFileSelectionMode(JFileChooser.FILES_AND_DIRECTORIES);
@@ -40,7 +61,5 @@ public class FolderBrowserDialog {
 			return folder;
 		}
 	}
-	
-}
 
-// TODO: search zip file
+}
